+++ chen_cppcguidll.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35013.160
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "chen_cppcguidll", "chen_cppcguidll\chen_cppcguidll.vcxproj", "{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "chen_cppcguilib", "chen_cppcguilib\chen_cppcguilib.vcxproj", "{A55AB0D1-A510-4DC8-A0AE-4D6BBCA5E9C5}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "sandbox", "sandbox\sandbox.vcxproj", "{80892FC0-412D-4713-9E2D-27A01E05AC85}"
@@ -17,14 +15,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Debug|x64.ActiveCfg = Debug|x64
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Debug|x64.Build.0 = Debug|x64
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Debug|x86.ActiveCfg = Debug|Win32
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Debug|x86.Build.0 = Debug|Win32
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Release|x64.ActiveCfg = Release|x64
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Release|x64.Build.0 = Release|x64
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Release|x86.ActiveCfg = Release|Win32
-		{A0A1F3BB-603C-4C89-AE06-33B8D662DA6B}.Release|x86.Build.0 = Release|Win32
 		{A55AB0D1-A510-4DC8-A0AE-4D6BBCA5E9C5}.Debug|x64.ActiveCfg = Debug|x64
 		{A55AB0D1-A510-4DC8-A0AE-4D6BBCA5E9C5}.Debug|x64.Build.0 = Debug|x64
 		{A55AB0D1-A510-4DC8-A0AE-4D6BBCA5E9C5}.Debug|x86.ActiveCfg = Debug|Win32
