--- chen_cppcguilib/basicthing.cpp
@@ -1,14 +1,0 @@
-#include "basicthing.h"
-
-basicthing::basicthing(string _inpData)
-{
-	data = _inpData;
-}
-basicthing::~basicthing()
-{
-
-}
-string basicthing::selfDraw()
-{
-	return data;
-}
