--- chen_cppcguilib/basicthing.h
@@ -1,14 +1,0 @@
-#pragma once
-#include<iostream>
-#include<string>
-using namespace std;
-
-class basicthing
-{
-public:
-	basicthing(string data);
-	~basicthing();
-	string selfDraw();
-private:
-	string data;
-};
