+++ chen_cppcguilib/include/cgui/cgui.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "components/basicImage.h"
+#include "components/basicProgressBar.h"
+#include "components/basicText.h"
+#include "components/multiLineText.h"
+#include "page.h"
